MIT-73: Update to Nibabel=5.*
The updated major release of Nibabel introduced an API change that lead
to ValueError being raised when the nibabel1Image() was called with
int64 data (with no header or dtype arguments).

Some of the instances in mrimagetools where nibabel1Image() were called
were modified to include the dtype argument matching the data.
Specifically, instances where data comes from outside of the module,
tests that produce int64 data inputs, and tests where it is cumbersome
to identify that the dtype of the data is not int64.

+++ src/mrimagetools/containers/test_image.py
@@ -95,7 +95,9 @@ def nifti_image_containers_a() -> list[NiftiImageContainer]:
 
     containers = []
     for image_type in [nib.Nifti1Image, nib.Nifti2Image]:
-        temp_img = image_type(np.zeros((1, 1, 1)), affine=np.eye(4))
+        temp_img = image_type(
+            np.zeros((1, 1, 1), dtype=np.int64), affine=np.eye(4), dtype=np.int64
+        )
         header = temp_img.header
 
         for key, value in NIFTI_HEADER.items():
